+++ framework/core/src/http/application/expressLikeApplication.ts
@@ -51,6 +51,26 @@ export abstract class ForklaunchExpressLikeApplication<
       openTelemetryCollector
     );
 
+    process.on('uncaughtException', (err) => {
+      this.openTelemetryCollector.log('error', 'Uncaught exception', {
+        err
+      });
+      process.exit(1);
+    });
+    process.on('unhandledRejection', (reason) => {
+      this.openTelemetryCollector.log('error', 'Unhandled rejection', {
+        reason
+      });
+      process.exit(1);
+    });
+    process.on('exit', () => {
+      this.openTelemetryCollector.log('info', 'Shutting down application');
+    });
+    process.on('SIGINT', () => {
+      this.openTelemetryCollector.log('info', 'Shutting down application');
+      process.exit(0);
+    });
+
     this.internal.use(createContext(this.schemaValidator) as RouterHandler);
     this.internal.use(cors(this.appOptions?.cors ?? {}) as RouterHandler);
   }