+++ framework/express/src/middleware/content.parse.middleware.ts
@@ -83,7 +83,7 @@ function contentParse<SV extends AnySchemaValidator>(options?: {
             });
 
             file.on('end', () => {
-              const fileBuffer = Buffer.concat(chunks);
+              const fileBuffer = Uint8Array.from(chunks.flat());
               body[fieldname] = fileBuffer.toString();
             });
           });