+++ framework/validator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @forklaunch/validator
 
+## 0.6.8
+
+### Patch Changes
+
+- package conflict resolution
+- fix minor buffer bugs and update subdependencies
+- Updated dependencies
+- Updated dependencies
+  - @forklaunch/common@0.3.8
+
 ## 0.6.7
 
 ### Patch Changes